refactor: remove components/utils folder
