+++ src/components/index.tsx
@@ -1,1 +1,2 @@
-export { Loader } from './loading';
+export { Loader } from "./loading";
+export * from "./HidePreventUnmount";
